Added normalize/unnormalize for tri patch params
Added methods to normalize and unnormalize the local parameterization
for triangular patch refinements. This is primarily for testing purposes
and is not final API.

Also, started updating glFVarViewer to support interpolation of triangle
patches for face-varying channels. Again, not ready to be merged but,
but a snapshot of some work in progress.